+++ web/src/nav/GlobalNavbar.tsx
@@ -86,6 +86,16 @@ export class GlobalNavbar extends React.PureComponent<Props, State> {
                 : '/.assets/img/sourcegraph-head-logo.svg'
         } else {
             logoSrc = '/.assets/img/sourcegraph-mark.svg'
+            const { branding } = window.context
+            if (branding) {
+                if (this.props.isLightTheme) {
+                    if (branding.light && branding.light.symbol) {
+                        logoSrc = branding.light.symbol
+                    }
+                } else if (branding.dark && branding.dark.symbol) {
+                    logoSrc = branding.dark.symbol
+                }
+            }
         }
 
         const logo = (